Isolate socket.io server
