+++ app/controllers/users.js
@@ -1,7 +1,7 @@
 const { wrap: async } = require('co');
 const jwt = require('jsonwebtoken');
 const mongoose = require('mongoose');
-const User = mongoose.model('User');
+const User = require('../models/user');
 const config = require('../config');
 
 /**
