+++ app/models/user.js
@@ -146,4 +146,4 @@ UserSchema.statics = {
   }
 }
 
-mongoose.model('User', UserSchema);
+module.exports = mongoose.model('User', UserSchema);
